pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "socrates"
version = "0.1.0"
description = "Local Semantic File Manager with Auto-Janitor"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Socrates Team"}
]
keywords = ["semantic-search", "file-manager", "vector-database", "offline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "watchdog>=3.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
socrates = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
